chore: ui 관련 패키지 정리 #17
