+++ app/src/main/java/com/ikseong/ucompass/MainActivity.kt
@@ -7,12 +7,9 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.rememberNavController
-import com.ikseong.ucompass.navigation.MainNavHost
+import com.ikseong.ucompass.ui.navigation.MainNavHost
 import com.ikseong.ucompass.ui.theme.UCompassTheme
 import dagger.hilt.android.AndroidEntryPoint
