--- app/src/main/java/com/ikseong/ucompass/create/viewmodel/CreateUiAction.kt
@@ -1,8 +1,0 @@
-package com.ikseong.ucompass.create.viewmodel
-
-sealed interface CreateUiAction {
-    data object OnCreateClick: CreateUiAction
-    data object OnConfirmClick : CreateUiAction
-    data object OnShareClick : CreateUiAction
-    data class UpdateTitleField(val text: String) : CreateUiAction
-}
