--- app/src/main/java/com/ikseong/ucompass/create/viewmodel/CreateUiEvent.kt
@@ -1,7 +1,0 @@
-package com.ikseong.ucompass.create.viewmodel
-
-sealed interface CreateUiEvent {
-    data object NavigateToFinish: CreateUiEvent
-    data object NavigateToHome : CreateUiEvent
-    data object ShareLink : CreateUiEvent
-}
