--- app/src/main/java/com/ikseong/ucompass/create/viewmodel/CreateUiState.kt
@@ -1,6 +1,0 @@
-package com.ikseong.ucompass.create.viewmodel
-
-data class CreateUiState(
-    val title: String = "",
-    val link: String = ""
-)
