+++ app/src/main/java/com/ikseong/ucompass/ui/create/viewmodel/CreateUiEvent.kt
@@ -1,0 +1,7 @@
+package com.ikseong.ucompass.ui.create.viewmodel
+
+sealed interface CreateUiEvent {
+    data object NavigateToFinish: CreateUiEvent
+    data object NavigateToHome : CreateUiEvent
+    data object ShareLink : CreateUiEvent
+}
