+++ app/src/main/java/com/ikseong/ucompass/ui/create/viewmodel/CreateUiState.kt
@@ -1,0 +1,6 @@
+package com.ikseong.ucompass.ui.create.viewmodel
+
+data class CreateUiState(
+    val title: String = "",
+    val link: String = ""
+)
