+++ app/src/main/java/com/ikseong/ucompass/ui/main/viewmodel/MainUiEvent.kt
@@ -1,0 +1,9 @@
+package com.ikseong.ucompass.ui.main.viewmodel
+
+sealed interface MainUiEvent {
+    data object RequestLocationPermission : MainUiEvent
+    data object EditProfileDialog : MainUiEvent
+    data class NavigateToRoom(val id: Long) : MainUiEvent
+    data object NavigateToCreateRoom : MainUiEvent
+    data object OpenGallery : MainUiEvent
+}
