+++ app/src/main/java/com/ikseong/ucompass/ui/navigation/Routes.kt
@@ -1,0 +1,16 @@
+package com.ikseong.ucompass.ui.navigation
+
+import kotlinx.serialization.Serializable
+
+sealed interface Routes {
+    @Serializable
+    data object Home: Routes
+
+    @Serializable
+    data object CreateRoomTitle : Routes
+    @Serializable
+    data object CreateRoomFinished : Routes
+
+    @Serializable
+    data class Room(val id: Long): Routes
+}
